requirements.txt
streamlit
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
Pillow
grad-cam
opencv-python-headless
numpy
albumentations==1.3.1
pandas
scikit-learn
fastapi
uvicorn[standard]
python-multipart
mlflow